% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_df}
\alias{spatial_df}
\title{Simulated Cluster-Level Data}
\format{
A data frame with columns:
\describe{
\item{unit}{Cluster identifier (character).}
\item{diversity}{Standardized diversity index.}
\item{ses}{Standardized socioeconomic status.}
}
}
\usage{
spatial_df
}
\description{
Cluster-level covariates corresponding to the spatial units in
\code{\link{indiv_df}}, used for the dispersion model in \code{\link[=bayes_badger]{bayes_badger()}}.
}
\keyword{datasets}
